1 fix bug save download file to wrong data source, 2 download large files, 3 clear sessions to save space

+++ src/Tagcade/Worker/Workers/ExecuteIntegrationJobWorker.php
@@ -100,7 +100,9 @@ class ExecuteIntegrationJobWorker
             fclose($fpIntegrationConfigFile);
 
             // remove temp file
-            unlink($integrationConfigFilePath);
+            if (is_file($integrationConfigFilePath)) {
+                unlink($integrationConfigFilePath);
+            }
         }
     }
